pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "django-template"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^4.1"
python = "^3.8"
django-model-utils = "^4.2.0"
tzdata = "^2022.6"
[tool.poetry.dev-dependencies]
